refactor: simplified package structure
