--- arve/data/__init__.py
@@ -1,3 +1,0 @@
-from data import Data
-
-__all__ = ["Data"]
