+++ arve/planets.py
@@ -1,5 +1,7 @@
 from typing import TypeVar
 
+import numpy as np
+
 from arve import ARVE
 
 TPlanets = TypeVar("TPlanets", bound="Planets")
@@ -10,7 +12,7 @@ class Planets:
 
     def __init__(self: TPlanets, arve: ARVE) -> None:
         self.arve = arve
-        self.parameters: dict = {}
+        self.parameters: dict[str, np.float64] = {}
 
     def add_planet(self: TPlanets) -> None:
         # TODO
