--- arve/planets/__init__.py
@@ -1,3 +1,0 @@
-from planets import Planets
-
-__all__ = ["Planets"]
