--- arve/star/__init__.py
@@ -1,3 +1,0 @@
-from star import Star
-
-__all__ = ["Star"]
